feat(widget): updated code
